requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
multidict==6.3.2
propcache==0.3.1
python-dotenv==1.0.1
python-multipart==0.0.7
requests==2.31.0
SQLAlchemy==2.0.27
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.19.0
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
openai==1.3.7 
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
frozendict==2.4.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.3.2
multitasking==0.0.11
numpy==2.2.4
openai==1.3.7
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2025.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.27
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
yfinance==0.2.55
